pyproject.toml
[project]
name = "x-api-bookmarks"
version = "0.1.0"
description = "X API ブックマーク・いいね取得テスト & AIラベリング"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
